% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ex}
\alias{data_ex}
\title{Univariable example dataset}
\format{
A data frame with 100 rows and 4 columns:
\describe{
\item{betaY}{Estimate of genetic association with the outcome}
\item{betaYse}{Standard error of estimate of genetic association with the outcome}
\item{betaX}{Estimate of genetic association with exposure}
\item{betaXse}{Standard error of estimate of genetic association with exposure}
}
}
\usage{
data_ex
}
\description{
Example dataset containing associations between 100 variants and a single exposure variable and an outcome
}
\keyword{datasets}
