% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_horse.R
\name{mvmr_horse}
\alias{mvmr_horse}
\title{Title}
\usage{
mvmr_horse(
  D,
  no_ini = 3,
  variable.names = "theta",
  n.iter = 10000,
  n.burnin = 10000
)
}
\arguments{
\item{n.burnin}{}
}
\description{
Title
}
