% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mv_ex}
\alias{data_mv_ex}
\title{Multivariable example dataset}
\format{
A data frame with 100 rows and 6 columns:
\describe{
\item{betaY}{Estimate of genetic association with the outcome}
\item{betaYse}{Standard error of estimate of genetic association with the outcome}
\item{betaX1}{Estimate of genetic association with exposure 1}
\item{betaX1se}{Standard error of estimate of genetic association with exposure 1}
\item{betaX2}{Estimate of genetic association with exposure 2}
\item{betaX2se}{Standard error of estimate of genetic association with exposure 2}
}
}
\usage{
data_mv_ex
}
\description{
Example dataset containing associations between 100 genetic instruments and two exposure variables and an outcome, as well as their corresponding standard errors.
This is taken from a multivariable simulation study, where 20\% of variants are pleiotropic, and pleiotropy is balanced.
}
\keyword{datasets}
