% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_horse.R
\name{mr_horse}
\alias{mr_horse}
\title{Univariable MR-Horse method}
\usage{
mr_horse(
  D,
  n.chains = 3,
  variable.names = "theta",
  n.iter = 10000,
  n.burnin = 10000,
  stan = FALSE,
  n.cores = parallelly::availableCores()
)
}
\arguments{
\item{D}{Data frame containing betaY, betaYse, betaX and betaXse, or MRInput object}

\item{n.chains}{Number of chains}

\item{variable.names}{Parameters to save estimates for}

\item{n.iter}{Number of iterations (not including warmup)}

\item{n.burnin}{Number of warmup iterations}

\item{stan}{Fit the model using stan, default is to use JAGS}

\item{n.cores}{Number of cores to use in parallel if running multiple chains, defaults to parallelly::availableCores()}
}
\value{
List
}
\description{
Causal effect estimation is performed in a Bayesian framework using a horseshoe shrinkage prior to account for both correlated and uncorrelated pleiotropic effects
}
\examples{
library(coda)

# Load example data
data(data_ex)

# Fit model with JAGS
MREx = mr_horse(data_ex)

# Check estimates
MREx$MR_Estimate
}
