% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_horse.R
\name{mr_horse}
\alias{mr_horse}
\title{Univariable MR-Horse method}
\usage{
mr_horse(
  D,
  n.chains = 3,
  variable.names = "theta",
  n.iter = 10000,
  n.burnin = 10000,
  stan = FALSE,
  n.cores = parallelly::availableCores(),
  fixed_tau = -1
)
}
\arguments{
\item{D}{data frame containing betaY, betaYse, betaX and betaXse, or MRInput object}

\item{n.chains}{number of chains}

\item{variable.names}{parameters to save estimates for, in addition to theta}

\item{n.iter}{number of iterations (not including warmup)}

\item{n.burnin}{number of warmup iterations}

\item{stan}{fit the model using stan, default is to use JAGS}

\item{n.cores}{number of cores to use in parallel if running multiple chains, defaults to parallelly::availableCores()}

\item{fixed_tau}{a fixed value for tau, otherwise tau will be estimated}
}
\value{
Output from the mr_horse() function is a list that contains:
$MR_Estimate: a data frame with the causal effect estimate (which is the posterior mean), standard deviation (i.e., the posterior standard deviation), upper and lower bounds of the 95\% credible interval, and the R-hat value                                                                                                               the posterior standard deviation), upper and lower bounds of the 95\% credible interval, and the R-hat value
}
\description{
Causal effect estimation is performed in a Bayesian framework using a horseshoe shrinkage prior to account for both correlated and uncorrelated pleiotropic effects
}
\examples{
# Fit model with JAGS
# data(data_ex)
# MREx = mr_horse(data_ex, n.burnin = 1000, n.iter = 2000)

# Check estimates
# MREx$MR_Estimate

# Fit model with Stan
# MREx = mr_horse(data_ex, n.warmup = 1000, n.iter = 2000, stan = TRUE)

# Save samples for tau parameter (in addition to theta)
# MREx = mr_horse(data_ex, n.warmup = 1000, n.iter = 2000, variable.names = 'tau')
# summary(MREx$MR_Coda)

# View diagnostic plots (trace and density)
# plot(MREx$MR_Coda[,c('theta','tau')])
#

}
