% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmr_horse.R
\name{mvmr_horse}
\alias{mvmr_horse}
\title{Title}
\usage{
mvmr_horse(
  D,
  no_ini = 3,
  variable.names = "theta",
  n.iter = 10000,
  n.burnin = 10000
)
}
\arguments{
\item{D}{Dataset containing betaY, betaYse and at least betaX1, betaX1se, betaX2 and betaX2se}

\item{no_ini}{Number of chains}

\item{variable.names}{Parameters to save estimates for}

\item{n.iter}{Number of iterations (not including warmup)}

\item{n.burnin}{Number of warmup iterations}
}
\description{
Title
}
