% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmr_horse.R
\name{mvmr_horse}
\alias{mvmr_horse}
\title{Multivariable MR-Horse method}
\usage{
mvmr_horse(
  D,
  n.chains = 3,
  variable.names = "theta",
  n.iter = 10000,
  n.burnin = 10000,
  n.cores = parallelly::availableCores()
)
}
\arguments{
\item{D}{Dataset containing betaY, betaYse and at least betaX1, betaX1se, betaX2 and betaX2se, or MRMVInput object}

\item{n.chains}{Number of chains}

\item{variable.names}{Parameters to save estimates for, in addition to theta}

\item{n.iter}{Number of iterations (not including warmup)}

\item{n.burnin}{Number of warmup iterations}

\item{n.cores}{Number of cores to use in parallel if running multiple chains, defaults to parallelly::availableCores()}
}
\value{
Output from the mr_horse() function is a list that contains:
$MR_Estimate: a data frame with the causal effect estimate (which is the posterior mean), standard deviation (i.e., the posterior standard deviation), upper and lower bounds of the 95\% credible interval, and the R-hat value                                                                                                               the posterior standard deviation), upper and lower bounds of the 95\% credible interval, and the R-hat value
}
\description{
Multivariable MR-Horse method
}
\examples{
# Fit model with JAGS
# MREx = mvmr_horse(data_mv_ex, n.warmup = 1000, n.iter = 2000)

# Check estimates
# MREx$MR_Estimate

# Fit model with Stan
# MREx = mvmr_horse(data_mv_ex, n.warmup = 1000, n.iter = 2000, stan = TRUE)
}
